--- js/create-dom-element.js
@@ -1,12 +1,0 @@
-const creatDOMElement = (markup, className) => {
-  let newElement;
-
-  newElement = document.createElement(`section`);
-
-  newElement.className = className;
-  newElement.innerHTML = markup;
-
-  return newElement;
-};
-
-export default creatDOMElement;
