+++ js/main.js
@@ -1,9 +1,8 @@
-import showPage from '../js/show-page.js';
 import {welcome} from '../js/welcome.js';
 
 /* Формирование стартовой страницы */
 const showWelcomePage = () => {
-  showPage(welcome);
+  return welcome().element;
 };
 
 showWelcomePage();
