--- js/show-page.js
@@ -1,10 +1,0 @@
-const app = document.querySelector(`.app`);
-
-/* Отображение определьной страницы */
-const showPage = ({page, init}) => {
-  app.replaceChild(page(), app.querySelector(`.main`));
-
-  init();
-};
-
-export default showPage;
