+++ js/welcome-view.js
@@ -1,0 +1,32 @@
+import AbstractView from '../js/abstract-view.js';
+import {showRandomPage} from '../js/show-random-page.js';
+
+export default class WelcomeView extends AbstractView {
+  constructor(className = `main main--welcome`) {
+    super();
+    this.className = className;
+  }
+
+  get template() {
+    return `
+      <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
+      <button class="main-play">Начать игру</button>
+      <h2 class="title main-title">Правила игры</h2>
+      <p class="text main-text">
+        Правила просты&nbsp;— за&nbsp;5 минут ответить на все вопросы.<br>
+        Ошибиться можно 3 раза.<br>
+        Удачи!
+      </p>
+    `;
+  }
+
+  bind() {
+    const onMainPlayClick = () => {
+      showRandomPage();
+    };
+
+    const mainPlay = this.app.querySelector(`.main-play`);
+
+    mainPlay.addEventListener(`click`, onMainPlayClick);
+  }
+}
